true's Reflow Controller PCB+Schematic
My version of the Reflow Controller.

This is REV2, untested. I have mostly tested REV1 so far. Yes, it is way
more complicated than the original. :)

- Schematic cleaned up and organized. Hopefully this is easier to read
now.
- Supports physically Arduino Pro Mini, Pro Micro or Nano V3.0
- No built in SSR's - this version is intended for 120V, and external
SSRs supporting the current needed for 1500W ovens are cheaper anyway.
Also allows for other options for output. This also means that the FAN
and HTR AC ports are just paralleled with AC IN; they don't need to be
used.
- As there are no built-in SSRs, the board is shorter by 2cm
- Header and connector pinouts labeled on board
- Header pinouts for LCD match the order most commonly found for this
LCD type, making connection easier
- No headers block the top or bottom of the Arduino header, so
programming / USB will not be blocked
- Supports using Nano V3.0's 3V3 output; no need for voltage regulator
(cheaper). Note that there is a low current limit for this but with the
level shifters and MAX31855's we are within this limit.
- Uses more common and cheaper optocoupler for the ZX circuit
- Series resistors for LCD comms (works with cheaper LCDs without this
built in)
- Level shifters for future SD card support