Re-added Timer3 for Micro, removed ARDUINO_TYPE
TimerThree was re-added since the Micro doesn't have Timer2. However,
TimerThree has a bug where it doesn't check to see if the processor is
capable, and either Arduino or the compiler is too stupid to see that
this isn't included anymore after the preprocessor, so either comment
the include out or modify TimerThree to check this.

Removed ARDUINO_TYPE and perform tasks depending on the type of
processor.