Optional support to store text strings in Flash
RAM likely won't run short on this project, but a macro has been added
wrapping all text, which as an example can be used as the Arduino F(x)
macro.

Also changed idleTemp to a defined value; this saved ~50bytes of Flash.

TFT_ defines are now LCD_ defines.

Need to look into why space is a consistent ~70 bytes higher when using
PROGMEM; the lack of variables in .bss should reduce this size, should
it not?