Using <avr/eeprom.h> over EEPROM.h
The code already makes use of <avr/eeprom.h> so why not use it
everywhere? Simplifies code a bit, and I would like to use eeprom_*_word
so here is the change. Saves ~40 bytes.

Changed the location of various variables to the end of EEPROM rather
than right after the profiles. If the profile count changes in firmware,
it won't put garbage in these variables anymore.