Simplified getItemValuePointer, rampRate now int16_t
getItemValuePointer was used to recall various values for display and
editing. It did so by updating indirect ptr for a double and an int16.
The function was changed instead to return pointer to type void, which
simplifies use and lets any time be used where necessary.

rampUpRate and rampDownRate are now int16_t and are stored as tenths of
a degree. The overhead for doing math to print these for editing is less
than the FP math necessary to work with them. int16_t is used as other
printing functions already use this type, so this saves space.

Additionally, lower bounds checking is in place for editing numbers - no
number in practice goes negative so it is simply a bounds check to zero.

Total savings of >250bytes. Code now fits on an Arduino Micro again.