timerIsr tick rate reduced from 10KHz to 4KHz
For this slow MCU, a 10KHz timer ISR is just too fast - context switch /
ISR to code execute time is pretty big on AVR so this ISR was stealing a
lot of time just being alive. From what I can tell changing this to 4KHz
won't really hurt anything but will improve responsiveness of the menus.
Save 6 bytes by changing when encoder is serviced in this interrupt.