Show temps on graph lines, related bugfixes
Added code to show temperatures on the graph lines, on the right-hand
side of the graph.

Wrote the number of bytes required to use each graph option alone.

Removed all use of Color565() and replaced with #define'd values. Saves
72 bytes.

Fixed bug in stop graph on done code that failed compilation of the
option was disabled.