Added menu area separator, neatened landscape view
Added a horizontal separator to separate the menu from the lower area. I
plan to do things in the lower area, and it also helps to visually
identify that there may be more than 5 things in the menu.

Darkened the bottom bar in the graph to separate the graph from the
heater/fan/rate status info.